test: split PDF equation script boundary coverage
